Deprecated the existing data source type, and moved the processor to a subfolder
The new data source will treat everything as observable streams, to make things more reactive and easier to react to changes.
