+++ source/components/cardContainer/dataSources/processor/dataSourceProcessor.service.tests.ts
@@ -5,8 +5,8 @@ import __object = services.object;
 import __transform = services.transform;
 
 import { DataSourceProcessor, IProcessResult } from './dataSourceProcessor.service';
-import { DataPager } from '../paging/index';
-import { ISort, SortDirection, Sorter, MergeSort } from '../sorts/index';
+import { DataPager } from '../../paging/index';
+import { ISort, SortDirection, Sorter, MergeSort } from '../../sorts/index';
 
 interface ITestObject {
 	value: number;
